requirements.txt
# Core Libraries
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0

# Parsing & NLP
python-docx==1.2.0
textstat==0.7.4
spacy==3.7.2
nltk==3.9.1

# Data Handling
numpy<2
pandas==2.3.3
openpyxl==3.1.5
lxml==6.0.2
et_xmlfile==2.0.0
pdfminer.six==20250506

# Networking & Environment
requests==2.32.5
python-dotenv==1.1.1
certifi==2025.10.5
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11

# Misc
typing_extensions==4.15.0
tzdata==2025.2
pytz==2025.2
six==1.17.0
cffi==2.0.0
pycparser==2.23
cryptography==46.0.3
blinker==1.9.0
MarkupSafe==3.0.3